补充大于15岁男生查询

+++ task3/src/com/jikexueyuan/caiwen/servlet/StudentController.java
@@ -13,23 +13,37 @@ import java.util.List;
 /**
  * Created by caiwen on 2016/6/11.
  */
-public class StudentController extends HttpServlet{
+public class StudentController extends HttpServlet {
 
     private StudentService studentService = new StudentService();
 
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException,
+            IOException {
         String action = req.getParameter("action");
+        List<Student> students;
         if (action == null) {
-            List<Student> students = studentService.getAllStudents();
+            students= studentService.getAllStudents();
             req.setAttribute("students", students);
             req.getRequestDispatcher("student/list.jsp").forward(req, resp);
-        } else if ("delete".equals(action)) {
-            String gender = req.getParameter("gender");
-            studentService.deleteAllMales(gender);
-        } else if ("add".equals(action)) {
-            studentService.add10Students();
-        }
+        } else
+            if ("get".equals(action)) {
+                String startAge = req.getParameter("startAge");
+                students = studentService.getStudents(startAge);
+                req.setAttribute("students", students);
+                req.getRequestDispatcher("student/list.jsp").forward(req, resp);
+            } else
+                if ("delete".equals(action))
+                {
+                    String gender = req.getParameter("gender");
+                    studentService.deleteAllMales(gender);
+                } else
+                    if ("add".equals(action))
+                    {
+                        studentService.add10Students();
+                    }
+
     }
+
 
 }
